Nowa logika odjazdu

+++ .vs/Projekt SO Pociag/kierownik.c
@@ -1,95 +1,3 @@
-/*#include "mojeFunkcje.h"
-
-//Obsługa sygnału SIGINT
-void handle_sigint(int sig) {
-    sem_destroy(sem_get(".", 2, 2));
-    destroy_message_queue(get_message_queue(".", 0));
-    destroy_message_queue(get_message_queue(".", 1));
-    exit(0);
-}
-
-
-
-int main() {
-    setbuf(stdout, NULL);
-    printf("\nNowy kierownik pociagu PID: %d",getpid());
-
-    //Inicjalizacja zmiennych
-    int max_passengers = 5;
-    int max_bikes = 2;
-    int passengers = 0;
-    int bikes = 0;
-    int train_ID;
-    int passanger_pid;
-    int timeout = 5;
-
-    //Przygotowanie wiaodmości i kolejek
-    struct message* entrance_message = malloc(sizeof(struct message));
-    int msq0 = get_message_queue(".", 0);
-    int msq1 = get_message_queue(".", 1);
-    struct message* train_message = malloc(sizeof(struct message));
-
-    //Przygotowanie semaforów
-    int entrance_sem = sem_create(".", 2, 2);
-    sem_set_value(entrance_sem, 0, 1);
-    sem_set_value(entrance_sem, 1, 1);
-
-    //Przygotowanie pamięci współdzielonej reprezentującej pociągi
-    //char* trains = shared_mem_attach(shared_mem_create(".",3,no_trains*(max_bikes+max_passengers)));
-
-    //Główna pętla kierownika
-    while (1) {
-        //Czeka na wjazd
-        receive_message(get_message_queue(".", 3), 1, train_message);
-
-        //Pętla ładowania pociągu
-        while (passengers != max_passengers)
-        {
-            if (passengers < max_passengers) {
-                if (receive_message_no_wait(msq0, 1, entrance_message)) {
-                    //Podnoszenie semafora dla pasażera bez roweru
-                    sem_raise(entrance_sem, 0);
-                    passengers++;
-                    passanger_pid = entrance_message->ktype;
-                    sleep(1);
-                    //printf("\nKierownik: pasazer %d wsiadl do pociagu", passanger_pid);
-                }
-            }
-            if (bikes < max_bikes && passengers < max_passengers) {
-                if (receive_message_no_wait(msq1, 1, entrance_message)) {
-                    //Podnoszenie semafora dla pasażera z rowerem
-                    sem_raise(entrance_sem, 1);
-                    passengers++;
-                    bikes++;
-                    passanger_pid = entrance_message->ktype;
-                    sleep(1);
-                    //printf("\nKierownik: pasazer %d wsiadl do pociagu z rowerem",  passanger_pid);
-                }
-            }
-            sleep(0.5);
-        }
-        //Informuje o pełnym pociągu
-        train_message->ktype = train_ID;
-        train_message->mtype = 1;
-        send_message(get_message_queue(".", 2), train_message);
-
-        //Czeka na wyjazd
-        receive_message(get_message_queue(".", 3), 1, train_message);
-
-        //Wyjazd
-        printf("\nPociag: Pociag %d odjechal", train_ID);
-        sleep(timeout);
-        printf("\nPociag: Pociag %d wrócił", train_ID);
-        train_message->ktype = train_ID;
-        train_message->mtype = 0;
-        send_message(get_message_queue(".", 2), train_message);
-
-        //Zerowanie liczników
-        passengers = 0;
-        bikes = 0;
-    }
-    return 0;
-}*/
 #include "mojeFunkcje.h"
 
 // Obsługa sygnału SIGINT
@@ -149,6 +57,7 @@ int main() {
                     train[passengers] = entrance_message->ktype;
                     passengers++;
                     passanger_pid = entrance_message->ktype;
+                    sleep(1);
                     printf("\nKierownik: pasażer %d wsiadł do pociągu %d.", passanger_pid, train_ID);
                 }
             }
@@ -159,9 +68,11 @@ int main() {
                     passengers++;
                     bikes++;
                     passanger_pid = entrance_message->ktype;
+                    sleep(1);
                     printf("\nKierownik: pasażer %d wsiadł do pociągu %d z rowerem.", passanger_pid, train_ID);
                 }
             }
+            sleep(0.5);
         }
 
         // Powiadomienie zawiadowcy, że pociąg jest pełny
@@ -181,7 +92,9 @@ int main() {
         passengers = 0;
         bikes = 0;
         for (int i = 0; i < max_passengers; i++){
-            kill(train[i], 2);
+            if (train[i] > 0 && train[i] != getpid()) {
+                kill(train[i], 2);
+            }
             train[i] = 0;
         }
         sleep(15);
